+++ include/raknet/FileListTransfer.h
@@ -84,10 +84,12 @@ public:
     /// \param[in] orderingChannel Passed to RakPeerInterface::Send()
     /// \param[in] _incrementalReadInterface If a file in \a fileList has no data, _incrementalReadInterface will be used to read the file in chunks of size \a chunkSize
     /// \param[in] _chunkSize How large of a block of a file to read/send at once. Large values use more memory but transfer slightly faster.
-    void Send(FileList *fileList, RakNet::RakPeerInterface *rakPeer, SystemAddress recipient, unsigned short setID, PacketPriority priority, char orderingChannel, IncrementalReadInterface *_incrementalReadInterface=0, unsigned int _chunkSize=262144*4*16);
+    void Send(FileList *fileList, RakNet::RakPeerInterface *rakPeer, SystemAddress recipient,
+              unsigned short setID, PacketPriority priority, char orderingChannel,
+              IncrementalReadInterface *_incrementalReadInterface=0, size_t _chunkSize=262144*4*16);
 
     /// Return number of files waiting to go out to a particular address
-    unsigned int GetPendingFilesToAddress(SystemAddress recipient);
+    size_t GetPendingFilesToAddress(SystemAddress recipient);
 
     /// \brief Stop a download.
     void CancelReceive(unsigned short setId);
